changed size_type and id_type to std::size_t

+++ include/gl/types/core.hpp
@@ -10,7 +10,7 @@
 
 namespace gl {
 
-using size_type = std::uint64_t;
+using size_type = std::size_t;
 using id_type = size_type;
 
 template <typename T>
